Initial work on Wasm/Browser
This does not work!

As of this commit the only thing that works is launching the program.
Much of this is not properly tested, but hey, at least it's someting!

Includes opus.a for testing purposes (Idk if it works).
Includes simpleserv python script to serve on port 8000 (Also for
testing purposes; Serves the output of dotnet publish).