Emscripten OpenAL; Remove opus.a
This commit makes use of the emscripten openal api!
Currently I'm doing some weird stuff with DllImport's to make it work.
(Idk if there is a better solution)

Also removed opus.a because that was what was causing compilation issues
the whole time! Which means that it's back to the drawing board for it.