Update to latest; Add NativeStorageService
Idk if I was supposed to do this in the rebase, but I fixed the broken
code by once again copying the Linux code and modifying it as needed...
Which actually just boils down to added the DllImport workaround and
that's it.

In fact it we could figure out an actual solution to it I guess we could
maybe do VoiceCraft.Client.Linux.*?

Also I think I did NativeStorageService correctly maybe. Idk about async
stuff though.